requirements.txt
python-dotenv
google-generativeai
langchain-community
langchain
streamlit
faiss-cpu
PyPDF2
-e .